requirements.txt
customtkinter
openai-whisper
ffmpeg-python
torch
torchaudio
